Get Gemfile working against JRuby
The following libs aren't compatible due to their respective C-bindings:
 - nokogiri
 - shotgun
 - thin
 - unicorn
 - yajl-ruby

Bumping sinatra to 1.4.2 as it has better java support than the version
we were previously using.

Set source to HTTPS rubygems instead of the old style of using the
symbol. This is due to how bundler in JRuby (latest) works. It throws a
warning if using the hold style.