Rough patch to test sorting facets by proportion.
Sorting organisation facets by number of documents matching each
organisation biases towards organisations with more content.  Since we
have orders of magnitude differences in the number of documents for each
organisation, this skews the top suggestions massively; BIS, for
example, very often comes top.

This patch experiments with addressing the problem by using the ratio of
the number of documents in the org matching the search, to the number of
documents in the org in total.  I've tried this with a few queries, and
it seems to give better results, but it needs more experimentation to
see if we want to use it.

One concern is that if we still display the count of documents in the
org to the user in the facet box, the order will be neither numeric nor
alphabetical, which may well confuse and annoy users.

This relates to story https://www.pivotaltracker.com/story/show/69915510

Also, this code needs refactoring and testing.