Encode pound signs as \u00a3
This isn't necessary in dev, or when running from the command line on a
preview box, but in the live preview environment, the pound signs were
causing a:

    Encoding::InvalidByteSequenceError: "\xC2" on US-ASCII

error when loading the schema json (in schema_config.rb, "load_json")

We're not sure why this is; current theories are that it's due to
something missing in locale data.  Rather than investigate right now,
we're just going to make the file be valid US-ASCII and move on.